Break cancellation tests into new module
